moved costing model to new file
